// types/feedback.ts
export interface FeedbackContent {
    satisfactoryResponse: string;
    timeToSignedNotes: string;
    totalDuration: string;
    responseAccuracy: string;
    decreaseInStress: string;
    overallExperience: string;
    feedback: string;
    userId: string | undefined;
    caseId: string | undefined;
}